pyproject.toml
[project]
name = "gregor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv",
    "openai",
    "supabase",
    "openai-agents>=0.2.2",
]
